requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
branca==0.7.1
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
Django==5.0.2
django-colorfield==0.11.0
django-debug-toolbar==4.3.0
django-geojson==4.0.0
django-leaflet==0.29.0
djangorestframework==3.14.0
folium==0.15.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
install==1.3.5
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
outcome==1.3.0.post0
pycparser==2.21
PySocks==1.7.1
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
rfc3986==1.5.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
wsproto==1.2.0
xyzservices==2023.10.1
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
branca==0.7.1
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
Django==5.0.2
django-colorfield==0.11.0
django-debug-toolbar==4.3.0
django-geojson==4.0.0
django-leaflet==0.29.0
djangorestframework==3.14.0
folium==0.15.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
install==1.3.5
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
outcome==1.3.0.post0
pycparser==2.21
PySocks==1.7.1
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
rfc3986==1.5.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
wsproto==1.2.0
tailwind==3.1.5 
django-tailwind==3.8.0
django_browser_reload==1.12.1
xyzservices==2023.10.1
